Refactor navigation and header structure across multiple pages; add agency page and update CSS guide
- Updated header navigation links in index.html, about.html, contact.html, and services.html to include "Agency" link and adjusted paths.
- Added header actions for signup and login buttons in relevant pages.
- Created new agency.html page with placeholder content and consistent header/footer structure.
- Archived main.js and related modules for future Phase 2 interactivity.
- Introduced CSS_MAIN_GUIDE.md to document CSS conventions and structure.
