+++ index.html
@@ -45,9 +45,13 @@
         <nav class="site-nav" aria-label="Main navigation">
           <a href="#jobs">Jobs</a>
           <a href="#companies">Companies</a>
+          <a href="pages/agency.html">Agency</a>
           <a href="pages/about.html">About</a>
-          <a href="pages/contact.html">Contact</a>
         </nav>
+        <div class="header-actions">
+          <a class="signup-link muted" href="#">Signup</a>
+          <a class="btn btn-primary login-btn" href="#">Log in</a>
+        </div>
       </div>
     </header>
 
@@ -279,7 +283,6 @@
       integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
       crossorigin="anonymous"
     ></script>
-    <script src="assets/js/main.js"></script>
     -->
   </body>
 </html>
