+++ pages/about.html
@@ -38,11 +38,15 @@
       <div class="container header-inner">
         <a class="brand" href="../">Japan SSW</a>
         <nav class="site-nav" aria-label="Main navigation">
-          <a href="../index.html">Home</a>
           <a href="../index.html#jobs">Jobs</a>
           <a href="../index.html#companies">Companies</a>
-          <a href="contact.html">Contact</a>
+          <a href="agency.html">Agency</a>
+          <a href="about.html">About</a>
         </nav>
+        <div class="header-actions">
+          <a class="signup-link muted" href="#">Signup</a>
+          <a class="btn btn-primary login-btn" href="#">Log in</a>
+        </div>
       </div>
     </header>
 
@@ -169,7 +173,6 @@
       integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
       crossorigin="anonymous"
     ></script>
-    <script src="../assets/js/main.js"></script>
     -->
   </body>
 </html>
