+++ pages/agency.html
@@ -1,0 +1,82 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width,initial-scale=1" />
+    <title>Agency — Japan SSW</title>
+    <link rel="stylesheet" href="/assets/css/main.css" />
+  </head>
+  <body>
+    <header class="site-header">
+      <div class="container header-inner">
+        <a class="brand" href="../">Japan SSW</a>
+        <nav class="site-nav" aria-label="Main navigation">
+          <a href="../index.html#jobs">Jobs</a>
+          <a href="../index.html#companies">Companies</a>
+          <a href="agency.html">Agency</a>
+          <a href="about.html">About</a>
+        </nav>
+        <div class="header-actions">
+          <a class="signup-link muted" href="#">Signup</a>
+          <a class="btn btn-primary login-btn" href="#">Log in</a>
+        </div>
+      </div>
+    </header>
+
+    <main class="section-padding">
+      <div class="container">
+        <h1>Agency (Placeholder)</h1>
+        <p class="muted">
+          This is a placeholder page for Agency. Replace with real content.
+        </p>
+
+        <section class="feature-card mt-md">
+          <h2>Registered Support Organizations</h2>
+          <p class="muted">
+            List of agencies and short descriptions will go here.
+          </p>
+        </section>
+
+        <p class="mt-lg">
+          <a class="btn btn-primary" href="/">Back to Home</a>
+        </p>
+      </div>
+    </main>
+
+    <footer class="site-footer site-footer--dark">
+      <div class="container footer-grid">
+        <div class="footer-cta">
+          <h3>Want to post jobs on JapanSSW?</h3>
+          <p class="muted">
+            Reach qualified candidates under the SSW programme.
+          </p>
+          <div class="footer-cta-actions">
+            <a class="btn btn-secondary" href="#">Employer Login</a>
+            <a class="btn btn-primary" href="#">Post a Job</a>
+          </div>
+        </div>
+
+        <div class="footer-links">
+          <div class="link-col">
+            <h4>Company</h4>
+            <ul>
+              <li><a href="../">Home</a></li>
+              <li><a href="about.html">About</a></li>
+              <li><a href="contact.html">Contact</a></li>
+            </ul>
+          </div>
+        </div>
+      </div>
+
+      <div class="container footer-legal">
+        <div class="legal-inner">
+          <small>&copy; 2025 Japan SSW. All rights reserved.</small>
+          <nav aria-label="Footer legal">
+            <a href="#">Terms</a>
+            <a href="#">Privacy</a>
+          </nav>
+        </div>
+      </div>
+    </footer>
+  </body>
+</html>
