+++ pages/contact.html
@@ -30,13 +30,17 @@
   <body>
     <header class="site-header">
       <div class="container header-inner">
-        <a class="brand" href="../index.html">Japan SSW</a>
+        <a class="brand" href="../">Japan SSW</a>
         <nav class="site-nav" aria-label="Main navigation">
-          <a href="../index.html">Home</a>
+          <a href="../index.html#jobs">Jobs</a>
+          <a href="../index.html#companies">Companies</a>
+          <a href="agency.html">Agency</a>
           <a href="about.html">About</a>
-          <a href="services.html">Services</a>
-          <a href="contact.html" aria-current="page">Contact</a>
         </nav>
+        <div class="header-actions">
+          <a class="signup-link muted" href="#">Signup</a>
+          <a class="btn btn-primary login-btn" href="#">Log in</a>
+        </div>
       </div>
     </header>
 
@@ -97,14 +101,12 @@
     </footer>
 
     <!-- Custom JS (if required) -->
-    <!-- <script src="../assets/js/main.js"></script> -->
     <!-- Phase 2 placeholder: Bootstrap JS bundle & site JavaScript
     <script
       src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
       integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
       crossorigin="anonymous"
     ></script>
-    <script src="../assets/js/main.js"></script>
     -->
   </body>
 </html>
