+++ pages/services.html
@@ -30,13 +30,17 @@
   <body>
     <header class="site-header">
       <div class="container header-inner">
-        <a class="brand" href="../index.html">Japan SSW</a>
+        <a class="brand" href="../">Japan SSW</a>
         <nav class="site-nav" aria-label="Main navigation">
-          <a href="../index.html">Home</a>
+          <a href="../index.html#jobs">Jobs</a>
+          <a href="../index.html#companies">Companies</a>
+          <a href="agency.html">Agency</a>
           <a href="about.html">About</a>
-          <a href="services.html" aria-current="page">Services</a>
-          <a href="contact.html">Contact</a>
         </nav>
+        <div class="header-actions">
+          <a class="signup-link muted" href="#">Signup</a>
+          <a class="btn btn-primary login-btn" href="#">Log in</a>
+        </div>
       </div>
     </header>
